src/main/java/ProductManager.java
public class ProductManager{

    ProductRepository repository;

    public ProductManager(ProductRepository repository) {
        this.repository = repository;
    }

    public void addProduct(Product product) {
        repository.addProduct(product);
    }


    //Сортируемый поиск
    public Product[] searchBy(String text) {
        Product[] result = new Product[0]; // тут будем хранить подошедшие запросу продукты
        for (Product product : repository.findAll()) {
            if (matches(product, text)) {
                Product[] add = new Product[result.length + 1];
                for (int i = 0; i < result.length; i++) {
                    add[i] = result[i];
                }
                add[add.length - 1] = product;
                result = add;
            }
        }
        return result;
    }

    // метод определения соответствия товара product запросу search
    public boolean matches(Product product, String search) {
        return product.getName().contains(search);
    }

}

